More cleanup - remove dead code, re-org stuff
