+++ app/src/main/java/com/dp/logcatapp/util/ActivityExt.kt
@@ -1,0 +1,12 @@
+package com.dp.logcatapp.util
+
+import android.app.Activity
+import android.view.WindowManager
+
+fun Activity.setKeepScreenOn(enabled: Boolean) {
+  if (enabled) {
+    window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+  } else {
+    window.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+  }
+}
