--- app/src/main/java/com/dp/logcatapp/util/ContextUtil.kt
@@ -1,13 +1,0 @@
-package com.dp.logcatapp.util
-
-import android.app.Activity
-import android.content.Context
-import android.content.ContextWrapper
-
-fun Context.findActivity(): Activity? {
-  return when (this) {
-    is Activity -> this
-    is ContextWrapper -> this.baseContext.findActivity()
-    else -> null
-  }
-}
