--- app/src/main/java/com/dp/logcatapp/util/HandlerCompat.kt
@@ -1,6 +1,0 @@
-package com.dp.logcatapp.util
-
-import android.os.Handler
-import android.os.Looper.getMainLooper
-
-fun mainHandler() = Handler(getMainLooper())
