+++ app/src/main/java/com/dp/logcatapp/util/StreamExt.kt
@@ -1,0 +1,19 @@
+package com.dp.logcatapp.util
+
+import java.io.IOException
+import java.io.InputStream
+import java.io.OutputStream
+
+fun InputStream.closeQuietly() {
+  try {
+    close()
+  } catch (_: IOException) {
+  }
+}
+
+fun OutputStream.closeQuietly() {
+  try {
+    close()
+  } catch (_: IOException) {
+  }
+}
