+++ app/src/main/java/com/dp/logcatapp/util/StringExt.kt
@@ -1,0 +1,8 @@
+package com.dp.logcatapp.util
+
+import android.annotation.SuppressLint
+import java.util.Locale
+
+@SuppressLint("DefaultLocale")
+fun String.containsIgnoreCase(other: String) =
+  lowercase(Locale.getDefault()).contains(other.lowercase(Locale.getDefault()))
