+++ app/src/main/java/com/dp/logcatapp/util/Utils.kt
@@ -1,7 +1,5 @@
 package com.dp.logcatapp.util
 
-import android.content.Context
-
 object Utils {
   fun bytesToString(bytes: Long): String {
     val units = arrayOf("B", "KB", "MB", "GB", "TB")
@@ -18,9 +16,4 @@ object Utils {
 
     return "%.2f %s".format(totalSize, unit)
   }
-
-  fun isUsingCustomSaveLocation(context: Context) =
-    context.getDefaultSharedPreferences().getString(
-      PreferenceKeys.Logcat.KEY_SAVE_LOCATION, ""
-    )!!.isNotEmpty()
 }
